% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{bctsGenSim}
\alias{bctsGenSim}
\title{Generate simulated data for the BC t-SNE algorithm}
\usage{
bctsGenSim(n = 200, p = 3000, k = 20, seed = NULL,
  categorical = TRUE)
}
\arguments{
\item{n}{integer of length 1, number of observations}

\item{p}{integer of length 1, number of parameters}

\item{k}{integer of length 1, number of components to define output, 'X'}

\item{seed}{integer of length 1, starting seed}

\item{categorical}{logical of length 1, switch simulation type, see details}
}
\value{
List with the following:
\itemize{
  \item X the (n x p) simulated matrix
  \item z the vector of group membership
  \item seed and categorical parameters stored as attributes 
}
}
\description{
Generate simulated data for the BC t-SNE algorithm
}
\details{
To add -- need to add data checks
}
\examples{
## Categorical example
s1 <- bctsGenSim(seed = 7231)
s1 <- bctsProcSim(s1)
p <- list(bty = "n", axes = FALSE, xlab = "", ylab = "", pch = 16)
m <- c("Unadjusted", "Pre-processed", "batch-corrected t-SNE")
par(mfrow = c(1, 3), oma = c(0, 0, 2, 0))
palette(c("#0072B2", "#E69F00", "gray60"))
mapply(plot, x = s1$embed, main = m, MoreArgs = c(col = list(s1$z), p))
title("Categorical group effect", outer = TRUE)
palette(value = "default")

## Continuous example
s2 <- bctsGenSim(seed = 7231, categorical = FALSE)
s2 <- bctsProcSim(s2)
s2col <- rainbow(21)[findInterval(s2$z, quantile(s2$z, seq(0, 1, 0.05)))]
par(mfrow = c(1, 3), oma = c(0, 0, 2, 0))
mapply(plot, x = s2$embed, main = m, MoreArgs = c(col = list(s2col), p))
title("Continuous group effect", outer = TRUE)

}
