% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bctsOG.R
\name{bctsOG}
\alias{bctsOG}
\title{Perform OG}
\usage{
bctsOG(X, Z, k = max(2, round(NCOL(X)/10)), rescale = FALSE)
}
\arguments{
\item{X}{numeric matrix}

\item{Z}{model matrix with covariate information}

\item{k}{integer of length 1}
}
\value{
numeric matrix (A + t(A))
}
\description{
Perform OG
}
\details{
Will take out when SOG gets added to CRAN
}
