% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{bctsProcEmbed}
\alias{bctsProcEmbed}
\title{Analyze the embeddings of simulated data}
\usage{
bctsProcEmbed(Y, z, n = 10, categorical = TRUE)
}
\arguments{
\item{Y}{numeric matrix with embeddings}

\item{z}{numeric vector of group membership}

\item{n}{interger of length 1, the number of nearest neighbors to analyze}

\item{categorical}{logical of length 1, switch simulation type, see details}
}
\value{
metric
}
\description{
Analyze the embeddings of simulated data
}
\details{
To add -- need to add data checks
}
