% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{bctsProcSim}
\alias{bctsProcSim}
\title{Process simulated data}
\usage{
bctsProcSim(s, k = 50)
}
\arguments{
\item{s}{list containing simulated data, returned by \link{\code{bctsGenSim}}}

\item{k}{integer of length 1, the number of initial dimensions to reduce X}
}
\value{
Modified sim object with the three embeddings stored in s$embed
}
\description{
Produce embeddings for the simulated data object
}
\details{
Need to add. -- need to add data checks to the function
}
